restructure modal component folder
